Aufteilung der Cards
